+++ resources/ReleaseNotes.md
@@ -1,5 +1,6 @@
 - Allow visualizing events in stackedplot
 - Customize display of "mag.Instrument" class
+- Move definition of `resample` method to `mag.TimeSeries` class
 - Move PNG export before FIG save
 - Do not restrict value of line to "double" only
 - Fix multiple small issues with HK view